app-docker/requirements.txt
mtcnn
fastapi[all]
python-multipart
starlette
jinja2
Pillow
requests
opencv-python-headless
numpy
tensorflow